Crate frost_rerandomized

FROST implementation supporting re-randomizable keys.

To sign with re-randomized FROST:

Do Round 1 the same way as regular FROST;

The Coordinator should generate a RandomizedParams and send the RandomizedParams::randomizer to all participants, using a confidential channel, along with the regular frost::SigningPackage;

Each participant should call sign and send the resulting frost::round2::SignatureShare back to the Coordinator;

The Coordinator should then call aggregate.

Re-exports

pub use frost_core;

Structs

RandomizedParams

Randomized parameters for a signing instance of randomized FROST.

Randomizer

A randomizer. A random scalar which is used to randomize the key.

Functions

aggregate

Re-randomized FROST signature share aggregation with the given RandomizedParams, which can be computed from the previously generated randomizer using RandomizedParams::from_randomizer.

sign

Re-randomized FROST signing using the given randomizer, which should be sent from the Coordinator using a confidential channel.