frost_rerandomized
Function sign

pub fn sign<C: Ciphersuite>(
    signing_package: &SigningPackage<C>,
    signer_nonces: &SigningNonces<C>,
    key_package: &KeyPackage<C>,
    randomizer: Randomizer<C>,
) -> Result<SignatureShare<C>, Error<C>>

Re-randomized FROST signing using the given randomizer, which should be sent from the Coordinator using a confidential channel.

See frost::round2::sign for documentation on the other parameters.